Move tracking components to their own namespace
After implementation of `PeriodTracker` from #8701, I've realized that component didn't really have a namespace to reside in, so I've looked up all components that share similarity with it and I've found out that `BreakTracker` and `IdleTracker` are best cases, so I've moved them all up to their own namespace `osu.Game.Tracking` and will add `PeriodTracker` to `osu.Game.Tracking.Components` as a sub-component useful for tracking components and other usages.
