+++ osu.Game.Tests/Visual/TestSceneOsuGame.cs
@@ -26,6 +26,7 @@ using osu.Game.Rulesets.Mods;
 using osu.Game.Scoring;
 using osu.Game.Screens.Menu;
 using osu.Game.Skinning;
+using osu.Game.Tracking;
 using osu.Game.Utils;
 using osuTK.Graphics;
