+++ osu.Game/Graphics/Containers/UserDimContainer.cs
@@ -7,7 +7,7 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Configuration;
-using osu.Game.Screens.Play;
+using osu.Game.Tracking;
 
 namespace osu.Game.Graphics.Containers
 {
