+++ osu.Game/Screens/Menu/ButtonSystem.cs
@@ -20,11 +20,11 @@ using osu.Framework.Platform;
 using osu.Framework.Threading;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
-using osu.Game.Input;
 using osu.Game.Input.Bindings;
 using osu.Game.Online.API;
 using osu.Game.Overlays;
 using osu.Game.Overlays.Notifications;
+using osu.Game.Tracking;
 using osuTK;
 using osuTK.Graphics;
 using osuTK.Input;
