+++ osu.Game/Screens/Play/PlayerLoader.cs
@@ -16,11 +16,11 @@ using osu.Framework.Threading;
 using osu.Game.Configuration;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
-using osu.Game.Input;
 using osu.Game.Overlays;
 using osu.Game.Overlays.Notifications;
 using osu.Game.Screens.Menu;
 using osu.Game.Screens.Play.PlayerSettings;
+using osu.Game.Tracking;
 using osu.Game.Users;
 using osuTK;
 using osuTK.Graphics;
